add new discount

+++ FAQPhone/FAQPhone/Models/DiscounModel.cs
@@ -11,9 +11,9 @@ namespace FAQPhone.Models
         public string _id { get; set; }
         public AccountModel owner { get; set; }
         public DepartmentModel category { get; set; }
-        public decimal price { get; set; }
+        public long price { get; set; }
         public int count { get; set; }
-        public decimal total { get; set; }
+        public long total { get; set; }
         public decimal used { get; set; }
         public AttributeModel type { get; set; }
         public DateTime beginDate { get; set; }
